Introduce `ManagedAccountTrait::add_utxos_from_transaction`
